Move constants to constants/ dir
