+++ constants/courses.ts
@@ -1,0 +1,6 @@
+// Courses with sections that may contains different descriptions
+// (Fundies vs Accelerated Fundies)
+export const COURSES_WITH_ALT_SECTIONS = [
+  { code: "CS", number: "2500" },
+  { code: "CS", number: "2510" },
+];
