+++ constants/paths.ts
@@ -1,0 +1,1 @@
+export const DATA_DIR_PATH = "./.data";
