+++ fetcher/fetch-descriptions.ts
@@ -1,5 +1,5 @@
 import { getCourseDescription } from "@/banner/description.ts";
-import { DATA_DIR_PATH } from "@/fetcher/constants.ts";
+import { DATA_DIR_PATH } from "@/constants/paths.ts";
 import { mayContainDifferentDescriptions } from "@/fetcher/util.ts";
 import { Course, Subject } from "@/types.ts";
 import { readJSON, writeJSON } from "@/util/file.ts";
