+++ fetcher/fetch-subjects.ts
@@ -1,5 +1,6 @@
 import { getAllSubjects } from "@/banner/subject.ts";
-import { DATA_DIR_PATH, TERMS } from "@/fetcher/constants.ts";
+import { DATA_DIR_PATH } from "@/constants/paths.ts";
+import { TERMS } from "@/constants/terms.ts";
 import { writeJSON } from "@/util/file.ts";
 
 const subjects = await getAllSubjects(TERMS);
