+++ fetcher/util.ts
@@ -1,10 +1,5 @@
-import { COURSES_WITH_ALT_SECTIONS } from "./constants.ts";
+import { COURSES_WITH_ALT_SECTIONS } from "@/constants/courses.ts";
 
-export const mayContainDifferentDescriptions = (course: {
-  code: string;
-  number: string;
-}) => {
-  return COURSES_WITH_ALT_SECTIONS.some(
-    (c) => c.code == course.code && c.number == course.number,
-  );
+export const mayContainDifferentDescriptions = (course: { code: string; number: string }) => {
+  return COURSES_WITH_ALT_SECTIONS.some((c) => c.code == course.code && c.number == course.number);
 };
